+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADConditionalAccessPolicy/MSFT_AADConditionalAccessPolicy.psm1
@@ -68,7 +68,7 @@ function Get-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $IncludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -82,7 +82,7 @@ function Get-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $ExcludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -173,6 +173,10 @@ function Get-TargetResource
         [Parameter()]
         [System.String]
         $CloudAppSecurityType,
+
+        [Parameter()]
+        [System.Boolean]
+        $SecureSignInSessionIsEnabled,
 
         [Parameter()]
         [System.Int32]
@@ -747,6 +751,7 @@ function Get-TargetResource
         CloudAppSecurityIsEnabled                = $false -or $Policy.SessionControls.CloudAppSecurity.IsEnabled
         #make false if undefined, true if true
         CloudAppSecurityType                     = [System.String]$Policy.SessionControls.CloudAppSecurity.CloudAppSecurityType
+        SecureSignInSessionIsEnabled             = $false -or $Policy.SessionControls.SecureSignInSession.IsEnabled
         #no translation needed, return empty string array if undefined
         SignInFrequencyIsEnabled                 = $false -or $Policy.SessionControls.SignInFrequency.IsEnabled
         #make false if undefined, true if true
@@ -851,7 +856,7 @@ function Set-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $IncludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -865,7 +870,7 @@ function Set-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $ExcludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -956,6 +961,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $CloudAppSecurityType,
+
+        [Parameter()]
+        [System.Boolean]
+        $SecureSignInSessionIsEnabled,
 
         [Parameter()]
         [System.Int32]
@@ -1848,30 +1857,42 @@ function Set-TargetResource
             $NewParameters.Add('grantControls', $GrantControls)
         }
 
-        if ($ApplicationEnforcedRestrictionsIsEnabled -or $CloudAppSecurityIsEnabled -or $SignInFrequencyIsEnabled -or $PersistentBrowserIsEnabled -or ($null -ne $DisableResilienceDefaultsIsEnabled))
+        if ($PSBoundParameters.ContainsKey('ApplicationEnforcedRestrictionsIsEnabled') -or $PSBoundParameters.ContainsKey('CloudAppSecurityIsEnabled') `
+            -or $PSBoundParameters.ContainsKey('SignInFrequencyIsEnabled') -or $PSBoundParameters.ContainsKey('PersistentBrowserIsEnabled') `
+            -or ($null -ne $DisableResilienceDefaultsIsEnabled) -or $PSBoundParameters.ContainsKey('SecureSignInSessionIsEnabled'))
         {
             Write-Verbose -Message 'Set-Targetresource: process session controls'
             $sessioncontrols = $null
             Write-Verbose -Message 'Set-Targetresource: create provision Session Control object'
-            $sessioncontrols = @{}
+            $sessioncontrols = @{
+                applicationEnforcedRestrictions = $null
+                cloudAppSecurity                = $null
+                secureSignInSession             = $null
+                signInFrequency                 = $null
+                persistentBrowser               = $null
+                disableResilienceDefaults       = $null
+            }
 
             if ($ApplicationEnforcedRestrictionsIsEnabled -eq $true)
             {
-                $sessioncontrols.Add('applicationEnforcedRestrictions', @{})
-                #create and provision ApplicationEnforcedRestrictions object if used
-                $sessioncontrols.applicationEnforcedRestrictions.Add('IsEnabled', $true)
+                $sessioncontrols.applicationEnforcedRestrictions = @{
+                    isEnabled = $ApplicationEnforcedRestrictionsIsEnabled
+                }
             }
             if ($CloudAppSecurityIsEnabled)
             {
                 $cloudAppSecurityValue = @{
-                    isEnabled            = $false
-                    cloudAppSecurityType = $null
+                    isEnabled            = $true
+                    cloudAppSecurityType = $CloudAppSecurityType
                 }
-
-                $sessioncontrols.Add('cloudAppSecurity', $CloudAppSecurityValue)
-                #create and provision CloudAppSecurity object if used
-                $sessioncontrols.cloudAppSecurity.isEnabled = $true
-                $sessioncontrols.cloudAppSecurity.cloudAppSecurityType = $CloudAppSecurityType
+                $sessioncontrols.cloudAppSecurity = $cloudAppSecurityValue
+            }
+            if ($SecureSignInSessionIsEnabled)
+            {
+                $secureSignInSessionValue = @{
+                    isEnabled = $SecureSignInSessionIsEnabled
+                }
+                $sessioncontrols.secureSignInSession = $secureSignInSessionValue
             }
             if ($SignInFrequencyIsEnabled)
             {
@@ -1882,7 +1903,7 @@ function Set-TargetResource
                     frequencyInterval = $null
                 }
 
-                $sessioncontrols.Add('signInFrequency', $SigninFrequencyProp)
+                $sessioncontrols.signInFrequency = $signinFrequencyProp
                 #create and provision SignInFrequency object if used
                 $sessioncontrols.signInFrequency.isEnabled = $true
                 if ($SignInFrequencyType -ne '')
@@ -1906,18 +1927,14 @@ function Set-TargetResource
             if ($PersistentBrowserIsEnabled)
             {
                 $persistentBrowserValue = @{
-                    isEnabled = $false
-                    mode      = $false
+                    isEnabled = $true
+                    mode      = $PersistentBrowserMode
                 }
-                $sessioncontrols.Add('persistentBrowser', $PersistentBrowserValue)
-                Write-Verbose -Message "Set-Targetresource: Persistent Browser settings defined: PersistentBrowserIsEnabled:$PersistentBrowserIsEnabled, PersistentBrowserMode:$PersistentBrowserMode"
-                #create and provision PersistentBrowser object if used
-                $sessioncontrols.persistentBrowser.isEnabled = $true
-                $sessioncontrols.persistentBrowser.mode = $PersistentBrowserMode
+                $sessioncontrols.persistentBrowser = $persistentBrowserValue
             }
-            if (-not [System.String]::IsNullOrEmpty($DisableResilienceDefaultsIsEnabled))
+            if ($DisableResilienceDefaultsIsEnabled)
             {
-                $sessioncontrols.Add('disableResilienceDefaults', $DisableResilienceDefaultsIsEnabled)
+                $sessioncontrols.disableResilienceDefaults = $DisableResilienceDefaultsIsEnabled
             }
             $NewParameters.Add('sessionControls', $sessioncontrols)
             #add SessionControls to the parameter list
@@ -2070,7 +2087,7 @@ function Test-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $IncludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -2084,7 +2101,7 @@ function Test-TargetResource
 
         [Parameter()]
         [System.String[]]
-        [validateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
+        [ValidateSet('none', 'internalGuest', 'b2bCollaborationGuest', 'b2bCollaborationMember', 'b2bDirectConnectUser', 'otherExternalUser', 'serviceProvider', 'unknownFutureValue')]
         $ExcludeGuestOrExternalUserTypes,
 
         [Parameter()]
@@ -2175,6 +2192,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $CloudAppSecurityType,
+
+        [Parameter()]
+        [System.Boolean]
+        $SecureSignInSessionIsEnabled,
 
         [Parameter()]
         [System.Int32]